updated public trans. code

+++ Data Extract/Code/pub_trans.R
@@ -10,7 +10,7 @@ pub_trans = get_acs(variables = "B08006_008",
                     geography="metropolitan statistical area/micropolitan statistical area") 
 
 pub_trans = pub_trans %>% rename("CBSA" = "GEOID", "Avg_User" = "estimate")
-pub_trans$CBSA = as.integer(pub_trans$CBSA)
+pub_trans$CBSA = as.numeric(pub_trans$CBSA)
 pub_trans = pub_trans[,c(1,2,4)]
